+++ templates/stores/index.html
@@ -1,10 +1,11 @@
 {% extends 'layouts/default.html' %} {% block main %}
 <h1>店家列表</h1>
-
+<!-- 會員顯示如下 -->
+{% if user.is_member %}
 <ul>
   {% for store in stores %}
   <li>
-    {% include "shared/favorite_btn.html" with store=store member=member %}
+    {% include "shared/favorite_btn.html" with store=store member=user.member %}
     <a href="{% url 'stores:show' store.id %}">
       <section>
         <h1>店家名稱：{{ store.name }}</h1>
@@ -18,9 +19,23 @@
   </li>
   {% empty %}
   <li>無商家資訊</li>
-  {% if user.is_store %}
-  <a href="{% url "stores:new" %}" class="btn">新增店家資訊</a>
-  {% endif %}
   {% endfor %}
 </ul>
+{% endif %}
+{% if user.is_store %}
+{% if store %}
+  <a href="{% url 'stores:show' store.id %}">
+    <section>
+      <h1>店家名稱：{{ store.name }}</h1>
+      <h1>店家地址：{{ store.address }}</h1>
+      <h1>店家電話：{{ store.phone_number }}</h1>
+      <h1>店家統編：{{ store.tax_id }}</h1>
+      <h1>開店時間：{{ store.opening_time }}</h1>
+      <h1>打烊時間：{{ store.closing_time }}</h1>
+    </section>
+  </a>
+{% else %}
+  <a href="{% url "stores:new" %}" class="btn">新增店家資訊</a>
+{% endif %}
+{% endif %}
 {% endblock %}
